BFF update after SSO changes

+++ internal/interfaces/repositories.go
@@ -6,8 +6,8 @@ import (
 
 type SsoRepository interface {
 	GetAllUsers() ([]*ssoentities.User, error)
-	GetUserByID(id int) (*ssoentities.User, error)
-	RegisterUser(userData ssoentities.RegisterUserDTO) (userID int, err error)
+	GetUserByID(id uint64) (*ssoentities.User, error)
+	RegisterUser(userData ssoentities.RegisterUserDTO) (userID uint64, err error)
 	LoginUser(userData ssoentities.LoginUserDTO) (*ssoentities.TokensDTO, error)
 	GetMe(accessToken string) (*ssoentities.User, error)
 	RefreshTokens(refreshTokensData ssoentities.TokensDTO) (*ssoentities.TokensDTO, error)
